Themes: move OBW API to Themes.php
