+++ client/dashboard/profile-wizard/steps/theme/uploader.js
@@ -50,7 +50,7 @@ class ThemeUploader extends Component {
 		const body = new window.FormData();
 		body.append( 'pluginzip', file );
 
-		return apiFetch( { path: '/wc-admin/themes', method: 'POST', body } )
+		return apiFetch( { path: '/wc-admin/themes/upload', method: 'POST', body } )
 			.then( ( response ) => {
 				onUploadComplete( response );
 				this.setState( { isUploading: false } );
